Improve platform endpoint patch
