+++ src/main/groovy/io/seqera/wave/service/pairing/PairingServiceImpl.groovy
@@ -54,20 +54,19 @@ class PairingServiceImpl implements PairingService {
 
     @Override
     PairingResponse acquirePairingKey(String service, String endpoint) {
-        final endpoint0 = patchPlatformEndpoint(endpoint)
-        final key = makeKey(service, endpoint0)
+        final key = makeKey(service, endpoint)
 
         def entry = store.get(key)
         if (!entry || entry.isExpired()) {
             final pairingId = LongRndKey.rndLong().toString()
-            log.debug "Pairing with service '${service}' at address $endpoint0 - pairing id: $pairingId (key: $key)"
+            log.debug "Pairing with service '${service}' at address $endpoint - pairing id: $pairingId (key: $key)"
             final keyPair = generate()
             final expiration = Instant.now() + lease
-            final newEntry = new PairingRecord(service, endpoint0, pairingId, keyPair.getPrivate().getEncoded(), keyPair.getPublic().getEncoded(), expiration)
+            final newEntry = new PairingRecord(service, endpoint, pairingId, keyPair.getPrivate().getEncoded(), keyPair.getPublic().getEncoded(), expiration)
             store.put(key,newEntry)
             entry = newEntry
         } else {
-            log.trace "Paired already with service '${service}' at address $endpoint0 - pairing id: $entry.pairingId (key: $key)"
+            log.trace "Paired already with service '${service}' at address $endpoint - pairing id: $entry.pairingId (key: $key)"
         }
 
         return new PairingResponse( pairingId: entry.pairingId, publicKey: entry.publicKey.encodeBase64() )
@@ -75,19 +74,8 @@ class PairingServiceImpl implements PairingService {
 
     @Override
     PairingRecord getPairingRecord(String service, String endpoint) {
-        final uid = makeKey(service, patchPlatformEndpoint(endpoint))
+        final uid = makeKey(service, endpoint)
         return store.get(uid)
-    }
-
-    static String patchPlatformEndpoint(String endpoint) {
-        // api.stage-tower.net --> api.cloud.stage-seqera.io
-        // api.tower.nf --> api.cloud.seqera.io
-        final result = endpoint
-                .replace('/api.stage-tower.net','/api.cloud.stage-seqera.io')
-                .replace('/api.tower.nf','/api.cloud.seqera.io')
-        if( result != endpoint )
-            log.debug "Patched Platform endpoint: '$endpoint' with '$result'"
-        return result
     }
 
     protected static String makeKey(String service, String endpoint) {
